The move's reason cites a record, like every other sentence
The Next-move card's reason was the one field the grounding filter let
through uncited. It was bounded by length and refused if it spelled a
record id, and nothing else — so free prose reached the reader
attributed as generated_by=model, whether it came from an instruction
embedded in a captured mail body on the deal's timeline or from a model
having a bad day. Every sibling sentence on the same card already had to
cite a record or be dropped.

The old reasoning was that the reason explains a move the RULES chose,
so what it rests on is the move's own evidence. That is true of the move
and not of the sentence.

move_reason is now a cited line held by the same keepGrounded filter,
and an ungrounded one is DROPPED rather than refusing the reply:
writtenMove keeps the deterministic reason the rules produced, so a card
grounded everywhere else still ships with a true sentence about the same
move. A reply answering the older shape — a bare string, from a
different provider or a cheaper lane — decodes through replyLines into
one uncited line and takes the same path, so the leniency that exists to
save a verdict cannot become a way past the check.

The filter moved to grounding.go on the way: model.go crossed the
500-line cap, and the two halves answer different questions — one asks
for prose, the other decides which of it is admissible.

This does NOT make the filter extractive. A sentence still has to cite a
record rather than quote one, so #2418's larger half stands.
